FAST_API/requirements.txt
mysql-connector
fastapi
uvicorn